refactor(plugins)!: move default configuration outside of lazy
