+++ lua/plugins/configs/Comment.lua
@@ -1,0 +1,1 @@
+return function(_, opts) require("Comment").setup(opts) end
