+++ lua/plugins/configs/aerial.lua
@@ -1,0 +1,1 @@
+return function(_, opts) require("aerial").setup(opts) end
