+++ lua/plugins/configs/alpha.lua
@@ -1,0 +1,1 @@
+return function(_, opts) require("alpha").setup(opts.opts) end
