+++ lua/plugins/configs/astrotheme.lua
@@ -1,0 +1,1 @@
+return function(_, opts) require("astrotheme").setup(opts) end
