+++ lua/plugins/configs/better_escape.lua
@@ -1,0 +1,1 @@
+return function(_, opts) require("better_escape").setup(opts) end
