+++ lua/plugins/configs/cmp.lua
@@ -1,0 +1,1 @@
+return function(_, opts) require("cmp").setup(opts) end
