+++ lua/plugins/configs/colorizer.lua
@@ -1,0 +1,1 @@
+return function(_, opts) require("colorizer").setup(opts) end
