+++ lua/plugins/configs/dressing.lua
@@ -1,0 +1,1 @@
+return function(_, opts) require("dressing").setup(opts) end
