+++ lua/plugins/configs/gitsigns.lua
@@ -1,0 +1,1 @@
+return function(_, opts) require("gitsigns").setup(opts) end
