+++ lua/plugins/configs/indent_blankline.lua
@@ -1,0 +1,1 @@
+return function(_, opts) require("indent_blankline").setup(opts) end
