+++ lua/plugins/configs/lspkind.lua
@@ -1,0 +1,4 @@
+return function(_, opts)
+  astronvim.lspkind = opts
+  require("lspkind").init(opts)
+end
