+++ lua/plugins/configs/luasnip.lua
@@ -1,0 +1,4 @@
+return function(_, opts)
+  if opts then require("luasnip").config.setup(opts) end
+  vim.tbl_map(function(type) require("luasnip.loaders.from_" .. type).lazy_load() end, { "vscode", "snipmate", "lua" })
+end
