+++ lua/plugins/configs/mason-lspconfig.lua
@@ -1,0 +1,6 @@
+return function(_, opts)
+  local mason_lspconfig = require "mason-lspconfig"
+  mason_lspconfig.setup(opts)
+  mason_lspconfig.setup_handlers { function(server) astronvim.lsp.setup(server) end }
+  astronvim.event "LspSetup"
+end
