+++ lua/plugins/configs/mason-null-ls.lua
@@ -1,0 +1,5 @@
+return function(_, opts)
+  local mason_null_ls = require "mason-null-ls"
+  mason_null_ls.setup(opts)
+  mason_null_ls.setup_handlers {}
+end
