+++ lua/plugins/configs/mason.lua
@@ -1,0 +1,1 @@
+return function(_, opts) require("mason").setup(opts) end
