+++ lua/plugins/configs/neo-tree.lua
@@ -1,0 +1,1 @@
+return function(_, opts) require("neo-tree").setup(opts) end
