+++ lua/plugins/configs/neodev.lua
@@ -1,0 +1,1 @@
+return function(_, opts) require("neodev").setup(opts) end
