+++ lua/plugins/configs/notify.lua
@@ -1,0 +1,5 @@
+return function(_, opts)
+  local notify = require "notify"
+  notify.setup(opts)
+  vim.notify = notify
+end
