+++ lua/plugins/configs/null-ls.lua
@@ -1,0 +1,1 @@
+return function(_, opts) require("null-ls").setup(opts) end
