+++ lua/plugins/configs/nvim-dap-ui.lua
@@ -1,0 +1,7 @@
+return function(_, opts)
+  local dap, dapui = require "dap", require "dapui"
+  dap.listeners.after.event_initialized["dapui_config"] = function() dapui.open() end
+  dap.listeners.before.event_terminated["dapui_config"] = function() dapui.close() end
+  dap.listeners.before.event_exited["dapui_config"] = function() dapui.close() end
+  dapui.setup(opts)
+end
