+++ lua/plugins/configs/nvim-treesitter.lua
@@ -1,0 +1,1 @@
+return function(_, opts) require("nvim-treesitter.configs").setup(opts) end
