+++ lua/plugins/configs/nvim-web-devicons.lua
@@ -1,0 +1,4 @@
+return function(_, opts)
+  require("nvim-web-devicons").set_default_icon(astronvim.get_icon "DefaultFile", "#6d8086", "66")
+  require("nvim-web-devicons").set_icon(opts)
+end
