+++ lua/plugins/configs/session_manager.lua
@@ -1,0 +1,1 @@
+return function(_, opts) require("session_manager").setup(opts) end
