+++ lua/plugins/configs/smart-splits.lua
@@ -1,0 +1,1 @@
+return function(_, opts) require("smart-splits").setup(opts) end
