+++ lua/plugins/configs/toggleterm.lua
@@ -1,0 +1,1 @@
+return function(_, opts) require("toggleterm").setup(opts) end
