+++ lua/plugins/configs/ufo.lua
@@ -1,0 +1,1 @@
+return function(_, opts) require("ufo").setup(opts) end
