+++ lua/plugins/configs/which-key.lua
@@ -1,0 +1,1 @@
+return function(_, opts) require("which-key").setup(opts) end
