+++ lua/plugins/configs/window-picker.lua
@@ -1,0 +1,1 @@
+return function(_, opts) require("window-picker").setup(opts) end
