+++ lua/plugins/git.lua
@@ -13,6 +13,5 @@ return {
       untracked = { text = "▎" },
     },
   },
-  default_config = function(opts) require("gitsigns").setup(opts) end,
-  config = function(plugin, opts) plugin.default_config(opts) end,
+  config = require "plugins.configs.gitsigns",
 }
