+++ lua/plugins/telescope.lua
@@ -38,12 +38,5 @@ return {
       },
     }
   end,
-  default_config = function(opts)
-    local telescope = require "telescope"
-    telescope.setup(opts)
-    astronvim.conditional_func(telescope.load_extension, pcall(require, "notify"), "notify")
-    astronvim.conditional_func(telescope.load_extension, pcall(require, "aerial"), "aerial")
-    astronvim.conditional_func(telescope.load_extension, astronvim.is_available "telescope-fzf-native.nvim", "fzf")
-  end,
-  config = function(plugin, opts) plugin.default_config(opts) end,
+  config = require "plugins.configs.telescope",
 }
