+++ lua/user_example/init.lua
@@ -185,7 +185,7 @@ local config = {
     -- {
     --   "L3MON4D3/LuaSnip",
     --   config = function(plugin, opts)
-    --     plugin.default_config(opts) -- include the default astronvim config that calls the setup call
+    --     require "plugins.configs.luasnip"(plugin, opts) -- include the default astronvim config that calls the setup call
     --     -- add more custom luasnip configuration such as filetype extend or custom snippets
     --     local luasnip = require "luasnip"
     --     luasnip.filetype_extend("javascript", { "javascriptreact" })
@@ -194,7 +194,7 @@ local config = {
     -- {
     --   "windwp/nvim-autopairs",
     --   config = function(plugin, opts)
-    --     plugin.default_config(opts) -- include the default astronvim config that calls the setup call
+    --     require "plugins.configs.nvim-autopairs"(plugin, opts) -- include the default astronvim config that calls the setup call
     --     -- add more custom autopairs configuration such as custom rules
     --     local npairs = require "nvim-autopairs"
     --     local Rule = require "nvim-autopairs.rule"
@@ -224,7 +224,7 @@ local config = {
     -- {
     --   "folke/which-key.nvim",
     --   config = function(plugin, opts)
-    --     plugin.default_config(opts)
+    --     require "plugins.configs.which-key"(plugin, opts)
     --     -- Add bindings which show up as group name
     --     local wk = require "which-key"
     --     wk.register({
